#381 refactor world methods to seperate etl work more
