+++ src/f10_world/world_tool.py
@@ -34,66 +34,3 @@ def get_all_brick_dataframes(dir: str) -> list[BrickFileRef]:
             valid_bricks.append(BrickFileRef(dir, filename, sheet_name, brick_number))
 
     return valid_bricks
-
-
-# from sqlalchemy import create_engine, Column, Integer, String, Date, Float
-# from sqlalchemy.orm import sessionmaker
-# from datetime import datetime
-
-
-# # 2. Create an engine and session
-# engine = create_engine("sqlite:///employees.db")
-# Base.metadata.create_all(engine)
-# Session = sessionmaker(bind=engine)
-# session = Session()
-
-# # 3. Read Excel data into a Pandas DataFrame
-# df = pd.read_excel("employees.xlsx")
-
-# # 4. Insert data from DataFrame to the database
-# for index, row in df.iterrows():
-#     employee = Employee(
-#         id=row["id"],
-#         name=row["name"],
-#         position=row["position"],
-#         hire_date=pd.to_datetime(row["hire_date"]).date(),
-#         salary=row["salary"],
-#     )
-#     session.add(employee)
-
-# # 5. Commit the session and verify data
-# session.commit()
-
-# # Optional: Print out the records
-# for employee in session.query(Employee).all():
-#     print(f"{employee.id}: {employee.name} - {employee.position} - {employee.salary}")
-
-# # 6. Close the session
-# session.close()
-
-
-def _create_events_agg_df(events_log_df: DataFrame) -> DataFrame:
-    events_agg_df = events_log_df[["face_id", "event_id"]].drop_duplicates()
-    events_agg_df["note"] = (
-        events_agg_df["event_id"]
-        .duplicated(keep=False)
-        .apply(lambda x: "invalid because of conflicting event_id" if x else "")
-    )
-    return events_agg_df.sort_values(["event_id", "face_id"])
-
-
-def get_pidgin_core_report_df(pidgin_dataframe: DataFrame) -> DataFrame:
-    pidgin_core_columns = [
-        "face_id",
-        "otx_wall",
-        "inx_wall",
-        "unknown_word",
-    ]
-    pidgin_core_df = DataFrame([], columns=pidgin_core_columns)
-    core_attrs = ["otx_wall", "inx_wall", "unknown_word"]
-    pidgin_agg_df = pidgin_dataframe.groupby("face_id").apply(
-        lambda x: {col: x[col].unique().tolist() for col in core_attrs}
-    )
-    pidgin_core_df = pidgin_agg_df.apply(PandaSeries)
-    pidgin_core_df = pidgin_core_df.reset_index()
-    return pidgin_core_df
